fix(user): remove mobile sales table, add scrollable-x sales, fix styling
