+++ src/components/TableUserSales/TableUserSales.tsx
@@ -1,8 +1,6 @@
 import React, { memo } from 'react';
 import { Action } from "../../types/entities/Action";
-import useWindowSize, { breakpoints } from "../../hooks/useWindowsSize";
 import { TableUserSalesDesktop } from "./TableUserSalesDesktop";
-import { TableUserSalesMobile } from "./TableUserSalesMobile";
 import { User } from '../../types/entities/User';
 
 interface TableUserSalesProps {
@@ -11,12 +9,8 @@ interface TableUserSalesProps {
   loading?: boolean,
 }
 const _TableUserSales = ({ user, sales, loading }: TableUserSalesProps) => {
-  const { width } = useWindowSize();
-  // todo infinite scroll
-  // todo virtualize table
-  return width !== undefined && width < breakpoints.sm ?
-    <TableUserSalesMobile sales={sales} loading={loading} />
-    : <TableUserSalesDesktop user={user} sales={sales} loading={loading} />
+  // todo virtualize table?
+  return <TableUserSalesDesktop user={user} sales={sales} loading={loading} />
 };
 
 export const TableUserSales = memo(_TableUserSales);
